moves parts of store.js to firefox-desktop.js
